% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i10_map_frame}
\alias{i10_map_frame}
\title{ICD10CM injury codes}
\format{
A data frame with 20,499 rows and 5 variables:
\describe{
  \item{dx}{A valid ICD-10 diagnosis code.}
  \item{cell}{The corresponding cell in the CDC Framework.}
  \item{PmCell}{The observed mortality for subjects with a diagnosis in this cell.}
  \item{PmCell}{The observed survival for subjects with a diagnosis in this cell.}
  \item{version}{The most recent date when this dataset was revised.  Encoded vyymmdd.}
  ...
}
}
\usage{
i10_map_frame
}
\description{
A dataset containing ICD-10 diagnosis codes and their properties
for assignment to the CDC Framework.
}
\keyword{datasets}
