% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{ICD10CM injury codes}
\format{
A data frame with 20,000 rows and 13 variables:
\describe{
  \item{temp_id}{A sequential number to identify individual subjects.}
  \item{died}{A binary indicator variable for death. 1 = died. 0 = survived.}
  \item{I10_DX1}{1st ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX2}{2nd ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX3}{3rd ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX4}{4th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX5}{5th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX6}{6th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX7}{7th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX8}{8th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX9}{9th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX10}{10th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX11}{11th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX12}{12th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX13}{13th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX14}{14th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX15}{15th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX16}{16th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX17}{17th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX18}{18th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX19}{19th ICD-10-CM injury code recorded on an encounter.}
  \item{I10_DX20}{20th ICD-10-CM injury code recorded on an encounter.}
  \item{Version}{The most recent date when this dataset was revised. Encoded vyymmdd.}
  ...
}
}
\usage{
testdata
}
\description{
A dataset containing a sample of trauma registry data for use in examples and tests.
}
\keyword{datasets}
