% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{i10_map_iciss}
\alias{i10_map_iciss}
\title{Table i10_map_iciss}
\format{
A data frame with 20,551 rows and 12 variables:
\describe{
  \item{dx}{A valid ICD-10 diagnosis code.}
  \item{digits1234} {The 4-digit version of dx.}
  \item{tot_int}{The number of cases in international data from which the observed survival was calculated.}
  \item{dsp_int}{The observed survival in international data for subjects with this diagnosis.}
  \item{dsp_int_c}{Same as above, but NA if tot_int < 5.}
  \item{tot_TQP}{The number of cases in TQP data from which the observed survival was calculated.}
  \item{dsp_TQP}{The observed survival in TQP data for subjects with this diagnosis.}
  \item{dsp_TQP_c}{Same as above, but NA if tot_TQP < 5.}
  \item{tot_NIS}{The number of cases in NIS data from which the observed survival was calculated.}
  \item{dsp_NIS}{The observed survival in NIS data for subjects with this diagnosis.}
  \item{dsp_NIS_c}{Same as above, but NA if tot_NIS < 5.}
  \item{Version}{The most recent date when this dataset was revised.  Encoded vyymmdd.}
  ...
}
}
\usage{
i10_map_iciss
}
\description{
A dataset containing ICD-10 diagnosis codes and their properties
for calculation of several versions of the ICD Injury Severity Score (ICISS).
}
\keyword{datasets}
