% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{ICD10CM injury codes}
\format{
A data frame with 20,000 rows and 13 variables:
\describe{
  \item{temp_id}{A sequential number to identify individual subjects.}
  \item{died}{A binary indicator variable for death. 1 = died. 0 = survived.}
  \item{dx1}{1st ICD-10-CM injury code recorded on an encounter.}
  \item{dx2}{2nd ICD-10-CM injury code recorded on an encounter.}
  \item{dx3}{3rd ICD-10-CM injury code recorded on an encounter.}
  \item{dx4}{4th ICD-10-CM injury code recorded on an encounter.}
  \item{dx5}{5th ICD-10-CM injury code recorded on an encounter.}
  \item{dx6}{6th ICD-10-CM injury code recorded on an encounter.}
  \item{dx7}{7th ICD-10-CM injury code recorded on an encounter.}
  \item{dx8}{8th ICD-10-CM injury code recorded on an encounter.}
  \item{dx9}{9th ICD-10-CM injury code recorded on an encounter.}
  \item{dx10}{10th ICD-10-CM injury code recorded on an encounter.}
  \item{dx11}{11th ICD-10-CM injury code recorded on an encounter.}
  \item{dx12}{12th ICD-10-CM injury code recorded on an encounter.}
  \item{dx13}{13th ICD-10-CM injury code recorded on an encounter.}
  \item{dx14}{14th ICD-10-CM injury code recorded on an encounter.}
  \item{dx15}{15th ICD-10-CM injury code recorded on an encounter.}
  \item{dx16}{16th ICD-10-CM injury code recorded on an encounter.}
  \item{dx17}{17th ICD-10-CM injury code recorded on an encounter.}
  \item{dx18}{18th ICD-10-CM injury code recorded on an encounter.}
  \item{dx19}{19th ICD-10-CM injury code recorded on an encounter.}
  \item{dx20}{20th ICD-10-CM injury code recorded on an encounter.}
  ...
}
}
\usage{
testdata
}
\description{
A dataset containing a sample of trauma registry data for use in examples and tests.
}
\keyword{datasets}
